+++ lib.commonjs/providers/provider-pocket.js
@@ -27,7 +27,7 @@ function getHost(name) {
             return "eth-goerli.gateway.pokt.network";
         case "matic":
             return "poly-mainnet.gateway.pokt.network";
-        case "maticmum":
+        case "matic-mumbai":
             return "polygon-mumbai-rpc.gateway.pokt.network";
     }
     (0, index_js_1.assertArgument)(false, "unsupported network", "network", name);