+++ lib.esm/providers/network.js
@@ -248,7 +248,8 @@ function injectCommonNetworks() {
             url: "https:/\/api.polygonscan.com/"
         }
     });
-    registerEth("maticMumbai", 80001, {
+    registerEth("matic-mumbai", 80001, {
+        altNames: ["maticMumbai", "maticmum"],
         //        priorityFee: 35000000000,
         etherscan: {
             //            apiKey: "W6T8DJW654GNTQ34EFEYYP3EZD9DD27CT7",