+++ types/providers/provider-etherscan.d.ts
@@ -13,7 +13,7 @@
  *  - Optimism (``optimism``)
  *  - Optimism Goerli Testnet (``optimism-goerli``)
  *  - Polygon (``matic``)
- *  - Polygon Mumbai Testnet (``maticmum``)
+ *  - Polygon Mumbai Testnet (``matic-mumbai``)
  *
  *  @_subsection api/providers/thirdparty:Etherscan  [providers-etherscan]
  */